Move rendered book to a versioned path.
Eventually, these notes are supposed to become a protocol specification, living
on a separate subdomain.  It's useful to version specs, so this adds a version
to the path, changing `penumbra.zone/XYZ` to `penumbra.zone/VERSION/XYZ`,
where `VERSION` is a branch name or git tag.

Unfortunately, the firebase cli tools want to deploy an entire site atomically,
rather than adding files, so currently, this deployment pipeline can't actually
handle tags; only the most recently deployed version will be available.
However, the URLs are in place, so the deployment pipeline can be fixed later
without redirects.

Finally, we add redirects immortalizing the current state of the notes, so that
existing links won't break, unless/until the underlying pages change.