Add tower-abci to the workspace.
This is a bit of a hack; tower-abci was developed in a separate repo.
Why include it in the workspace at all? The main reason is to have it
included in the auto-rendered docs.  However, git submodules aren't
super easy to work with, so in the long run this might not be a great
solution. Two possible alternatives:

- merge `tower-abci` into this repo, replaying git history;
- change the doc building process to have both.