crypto: create new crate, scratch work
Adds asset identifiers and value commitment generator computation.
Adds a basic test of value commitments with different bases.

This commit changes the workspace to use a Penumbra fork of some of the
arkworks crates.  The current change is in the Display impl for Fp* instances.
The fork's `ours` branch is based off of `v0.3.0`, in order to not pick up any
unreleased breaking changes.