crypto: rearrange {Extended,}SpendingKey to Spend{Seed, Key}
This commit rearranges the `SpendingKey` and `ExtendedSpendingKey` to new
types, called `SpendSeed` and `SpendKey`.

`SpendSeed` corresponds to the Zcash "spending key"; it's the root key
material, and is just a refinement type on an array of 32 bytes.

`SpendKey` corresponds to the Zcash "expanded spending key"; it's renamed to
reflect the fact that it is the primary spending key, and does not represent a
separation of capability from the `SpendSeed`.

Next, the implementation is restructured so that most of the key heirarchy is
derived at once and cached inside the `SpendKey`.  For instance, rather than
have a separate method to derive the ovk or nsk from the spend seed, it can be
done just within the `SpendKey` constructor.