refactor: `Note` no longer takes `Address`
When we do the note commitment integrity check as part of the
output proof verification, we create a note based on the g_d and
pk_d provided in the auxiliary inputs.

The problem is that note creation previously required a full `Address`
which has other fields that it would be nice not to include in the
output proof (namely the ClueKey).

So instead, we just hold g_k and pk_d.