refactor: `Note`  (#98)
* refactor: store Fq, ka::Public, diversifier on Note

Replaces the diversified generator on the Note with the diversifier
instead. There's now a method on the note to access the diversified
generator point.

In addition, we pass in the transmission key as `ka::Public`, and store
this in addition to the s-component as an `Fq`. The reason we neeed both
is that the `Fq` is used when creating note commitments, and the
`ka::Public` is used on the note when checking the diversified address
integrity as part of the `SpendProof` verification.

* make all fields private on Note