proto: add stub of Wallet service protos.
Starts work on #33.

- [x] defines a `.proto` specification of the client protocol;
- [x] adds support for generating Rust proto types.

Instead of generating the protos using `prost` directly, the wallet protos are
generated using `tonic`, but using the same `prost` config as the other
protos. The reason is that we define a GRPC service for the wallet protocol,
and we want to derive the implementation using `tonic`, but this requires using the protos.  The alternative would be to somehow expose the protos to other crates, and use `tonic` there, but that's not as nice.

For the moment, the stub protocol places full trust in the endpoint for
integrity and privacy, but engineering away this trust is a future design
goal.

Currently, there are only two RPC methods:

- `CompactBlockRange` returns a stream of `CompactBlocks`, containing a height
  and a list of `StateFragments`, which have just the information needed to do
client-side scanning.

- `TransactionByNote` allows querying for a full transaction body by note
  commitment.  This allows a client to get full information on a transaction
once they've identified it through scanning.  This is quite bad for privacy if
the server weren't trusted, as it precisely identifies the notes the client is
interested in. In the future, it should be replaced with something that leaks
less information.