crypto: use nonce of all 1s for `Memo` encryption
For Note encryption we have a key that is based on `epk`, the ephemeral
public key, and a shared secret. Now, the `Memo` encryption also uses
the same key, so it's important not to use the same
all 0 nonce (all 0 nonce from the ZCash spec) for our `Memo` encryption.
Instead, we use all 1s. .