feat: Note, memo, ovk encryption/decryption (#107)
* feat: `Note` encryption

* feat: OVK encryption

* feat: memo encryption

* use blake2b-256 everywhere as a KDF

* refactor: encrypt note plaintext as Vec<u8>

* feat: return errors, using thiserror if enum existed, else anyhow

* feat: `Note` decryption

* feat: memo decryption

* crypto: use nonce of all 1s for `Memo` encryption

For Note encryption we have a key that is based on `epk`, the ephemeral
public key, and a shared secret. Now, the `Memo` encryption also uses
the same key, so it's important not to use the same
all 0 nonce (all 0 nonce from the ZCash spec) for our `Memo` encryption.
Instead, we use all 1s. .

* use [1,0,..] instead of [1,1,..] for memo nonce

* address comments from review