pd: avoid use of block_on from inside of tokio
We can't use block_on inside the body of Service::call, since it will block the
whole thread until the future completes, but the future might be waiting on the
thread.  However, we still want to be able to have certain requests processed
sequentially.  To do this, we add a "completion tracker" that allows the
service to return Poll::Pending until the request future has been driven to
completion.  Thanks to @hawkw for describing the shape of the solution.

Co-Authored-By: Eliza Weisman <eliza@buoyant.io>