feat(pd): `mempool_nullifiers` is `<Arc<Mutex<BTreeSet<...>>>`
In order to inspect and modify `mempool_nullifiers` after performing
a database query, it needs to be within an `async` block in `CheckTx`.

This means that we need to modify the type of `mempool_nullifiers`
such that we can clone a reference to `self.mempool_nullifiers` and
move it into the future.