pcli: add file-backed ClientState wrapper
This changes the format from bincode to JSON, for no principled reason other
than that it's neat to be able to look at the fields in an editor.

The ClientStateFile wrapper in `pcli` holds a `penumbra_wallet::ClientState`
and a `PathBuf`, and uses `Deref`/`DerefMut` to present as a `ClientState`.
However, it also has an extra `commit` method that writes the state to disk.