pd: temporary workaround for fixed NCT issue
This stores a list of recent NCTs instead of just checking against the most
recent one.  However, it doesn't load this from the database on launch, so it's
possible for a new node to disagree with existing nodes about transaction
validity if it launched recently.  This isn't a problem for us at the moment,
since we're only running one node, but this implementation isn't correct
behavior and should be at least fixed to be consistent across launch times.
Fixing this for real could be combined with streamlining our DB accesses as we
process transactions.