pd: rework and rename `create-genesis` to `create-genesis-template`
Our create-genesis subcommand was serving two roles:

1. using command line parameters to generate JSON output that could be put into
the `app_state` field of the Tendermint `genesis.json` file;
2. constructing genesis notes given an input of genesis allocations.

The second role was only necessary because we made the `app_state` record
notes, rather than recording allocations directly, even though they have the
same information.  Instead, we can record the allocations as part of the
genesis state, and compute the notes as needed.  (Digging into this revealed
that the genesis transaction was being created non-deterministically, which
would cause an instant consensus divergence on a multi-node system; that's
fixed along the way).

The first role is not tenable as the scope of the Penumbra configuration
increases.  For instance, we now need to specify details about the validator
set.  Currently, it's not possible to feed the output of `create-genesis`
directly into the `genesis.json` file, because not all the fields are included.
This makes running nodes difficult.

Adding more and more command line options to the tool that generates the
config data, and then having people record those options in a script or
something, is much less ergonomic than just editing the config data directly.

To edit the config data directly, though, it's helpful to have a complete
example of the format.  So, the `create-genesis` command is renamed to
`create-genesis-template`, and it generates an editable example file.

Now that we specify validator info in the `app_state` together with the genesis
allocations, we'll need to change our setup for deployments so that the
`genesis.json` file is specified in advance.  This means that we'll need to
change the setup so that validator identity keys are chosen in advance (and
written into the genesis file).