pd: avoid deadlock by refactoring `CompletionTracker` -> `Sequencer`
This change attempts to fix an issue observed on the testnet where the node
stalled.  Investigating revealed that the response future from a `DeliverTx`
message, which is supposed to be processed sequentially, was not being polled.
(perhaps because the thing that would be polling it was waiting on sending
another request first?).

This caused the `App` to not process any *other* requests, as it (correctly)
signaled unreadiness. The root of this problem is that, while the `App`
requires that the request should be processed sequentially, it doesn't actually
take responsibility for doing so (by spawning a new task that would drive the
processing to completion).

This change avoids this failure mode, by changing the `CompletionTracker` that
tracks someone else's execution of a future (and relies on its caller to
correctly handle signaling completion) into a `Sequencer` that drives execution
of a future and does book-keeping internally.