Add protos for staking messages, set up code infrastructure
This chunk of work was meant to be simple and turned into an unexpectedly deep
rabbithole, but going down the rabbithole revealed some lessons for our
development process and helped build out a bunch of basic infrastructure.

The start was the creation of `stake.proto`, meant to define some stub proto
messages that we could use to do staking and delegation actions:

- defining (or updating) a validator;
- delegating to a validator;
- undelegating from a validator.

Delegation and undelegation messages are fairly straightforward: they specify
the validator, the unbonded amount, and the epoch.  (Those details determine
the rest of the parameters; the implementation work is in the checks that
they're consistent with the rest of the transaction, rather than in the data in
the message).

To avoid a separate process for creating a new validator and updating its
configuration, the idea is to use the same message with create-or-update
semantics.  This configuration data should be signed by the validator, and, as
with the other transaction structures, we split out the data-to-be-signed into
a separate message (`Validator`), and have the signed message
(`ValidatorDescription`) consist of a message+signature pair.  A transaction
can include a `ValidatorDescription` message to (re)define a validator.

Because this message includes a signature from the validator it configures,
it's self-authenticating.  However, a third party could create a transaction
including an authentic but stale validator description to roll back a
validator's configuration to some previous state.  To prevent this, the
`Validator` message includes a sequence number, so that the chain can require
that redefinitions have increasing sequence numbers, forcing each update to
pass through a signing oracle.

To specify the contents of the `Validator` description itself, we look at
the contents of the Cosmos SDK `create-validator` transaction as a model [0].
Some fields are just basic metadata, like the name, website, and description,
and others we ignore:

- initial commission rate, since we use funding streams;

- maximum commission and commission max change rate, which add implementation
  complexity to prevent a scenario with low impact (a validator advertises 0%
commission, then updates it to 100% in the hope the delegators won't notice)
that could be addressed by other means (social pressure, easy redelegation);

- minimum self-delegation, which doesn't fit cleanly with our delegation token
  model (and has unclear impact in a system with native delegation tokens
anyways).

However, looking at the address and pubkey fields reveals a problem with our
existing data modeling: Cosmos uses separate keys for the validator's identity
(determining the address, etc), and Tendermint consensus.  This is important
for various reasons, but most importantly, it allows the validator's identity
key to be stored offline, unlike the consensus key, which must be used online
to sign blocks.

The new `Validator` message distinguishes between the validator's
identity key (a decaf377-rdsa spend authorization key) and the validator's
consensus key (a Tendermint Ed25519 pubkey).  The validator's identity key
authenticates the consensus key by signing the `Validator` message.

Our existing code didn't notice this distinction, and changing the code to
account for it required propagating a bunch of other changes to the state
handling, effectively rewriting it.  While rewriting code isn't the end of the
world (especially when going from a first-pass of mapping out functionality to
a second-pass, coherent organization), in this instance the process was less
than ideal, and it would be good to figure out how we could try to improve it
in the future.

One observation is that it's not that the existing code was bad, but that we
worked "upwards" from the Tendermint API towards some idea of higher-level
application requirements, rather than specifying the eventual requirements in
more detail first, and then working from those requirements to the
implementation.  Writing down the requirements (in the form of proto
definitions for the messages the system uses to communicate) revealed a bunch
of details that we didn't previously think of.  The identity/consensus
distinction is a big one, but there are other, more trivial ones, like the fact
that we should record a validator's name and website, or that we'll need to
record both the exchange rates and reward rates.

This observation points to a potentially improved process: for each chunk of
functionality,

1. Write protos that define the data structures it will use;
2. Write domain types corresponding to those protos, and write the data
validation code as part of the conversions bundled with our `Protobuf` trait;
3. Implement the functionality using the domain types.

This way, in step 1, we can see how the data structures will fit together in
code review and map out the functionality without having to worry about the
details of the implementation, and the different components of step 3 can be
more easily parallelized, because people can work on different chunks of
implementation work using already-defined data structures.

Finally, another side benefit of centering the proto structures is that it
provides a clear definition of the data format, a centralized pathway for
validation (the conversion between the proto types and the domain types), and a
single location to define serializations in other formats (e.g., JSON).  Rather
than using Serde to derive serialization implementations ad-hoc, making every
structure definition potentially load-bearing and littering the code with
custom format-shaping annotations, we can use Serde on the proto structures.
This commit sets up the infrastructure to do that, based on the approach used
in `tendermint-rs`:

- The `penumbra-proto` `build.rs` has machinery to inject Serde derives, or
  format-shaping annotations ("serialize as hex string", "serialize as base64
string", etc.) into the Rust types generated from the proto files;

- The domain type for some proto type has a `TryFrom` impl that performs all of
  the validation logic;

- The domain type `Foo` derives `Serialize, Deserialize` with
  `#[serde(try_from = "pb::Foo", into = "pb::Foo")]`, so that the domain type's
Serde implementations are determined by the proto type, and there's a common
data validation pathway for both Protobuf and Serde serialization.

[0]: https://hub.cosmos.network/main/validators/validator-faq.html#how-to-become-a-validator