crypto: make sign of value commitments consistent
Previously, we treated the value commitment in an output description as a
commitment to the (positive) value accruing to the output notes, rather than as
a commitment to the (negative) value contributed (i.e., withdrawn) from the
transaction's value balance.  This meant that when summing the value
commitments for each action, we needed to handle spends and outputs
differently, because they had inconsistent signs.

Instead, this commit changes the value commitment of the output description to
commit to the action's contribution to the transaction's value balance (a
negative number).  This means that the semantics of the action (whether it is
contributing or withdrawing) are encapsulated in the action logic, rather than
leaking out to the code that handles the actions.