pd: make next epoch's rates available
Staking actions are supposed to make use of the rates in the epoch in which
they take effect (the next epoch), so we need to compute them in advance. This
commit changes the end-epoch processing to more cleanly model the epoch
transition (it happens after the end of the last block in the epoch) and
precompute the next epoch's rates.