pcli: factor out network constructors into methods on Opt
Adding these methods means that they can be shared across all client
functionality.

Adding these methods to Opt means we don't have to pass around the specific URI
fields, and opens the door to refactoring Opt into App.

This also allows us to avoid committing changes involving pending note tracking
to the state when the transaction is rejected.  Doing this requires removing
the `.commit()` call from the `ClientStateFile`'s `Drop` impl, because
otherwise we have no way to decide whether or not to commit changes to the
state (they would always be committed, which is not what we want).