pd: fix nullifier handling bug
The nullifier lookup-or-insert step was being performed *before* we performed
stateful transaction verification.  This meant that if a transaction was
submitted to the mempool that passed stateless verification but failed stateful
verification, any notes that it had touched would become unspendable, because
they'd already been added to the mempool nullifier set, even though the
transaction was rejected and the notes were not spent.  Instead, perform
nullifier checks last, following stateful verification.