feat(metrics): add revealed nullifier counter
The metrics need to persist between node restarts, so they need
to update from the content in the database. Here we add the spent
nullifier counter in BeginBlock by counting the nullifier rows in
the nullifiers table.

To update the counter, we want to update its absolute value, which
we can do using a feature added in metrics 0.18.0 - the absolute_counter!
macro.