pd: use `tokio_util::sync::PollSender` in consensus svc
In v0.7 of `tokio-util`, the `PollSender` API was changed to have the
semantics that most users expect (see tokio-rs/tokio#4214). The
`poll_send_done` method was replaced with a `poll_reserve` method, and
the implementation rewritten to drive a `Sender::reserve_owned` future
that's consumed in `send_item`. Now, `PollSender` essentially implements
exactly the same code that was written by hand in the consensus service.
We can simplify the consensus service significantly by upgrading the
`tokio-util` dependency to 0.7 and replacing the hand-written version
with `PollSender`.

There should be no functional change as a result of this refactor.