Add a `PenumbraStore` extension trait for accessing common data.
This trait is motivated by the need to have multiple application components
that access common data.  So far we've had the idea that there are going to be
sub-trees of the JMT for the different application areas, like shielded_pool/,
ibc/, staking/, etc, and that each component is responsible for managing the
state in there, and that we handle "transactions affecting multiple areas" by
"each area does its part separately".

However, there's going to be some cases where we'd want to have different
components accessing common data -- like the epoch_duration.  In individual
components, most of the actual "business logic" of the component is separated
from accesses to the state store, by means of accessor methods that just handle
interfacing with the tree.

This is obviously a good approach, since we don't have to deal with the store
paths or whatever except in the get/put methods themselves, but how does that
work for the epoch_duration, which we may want to access in multiple
components?

One solution is sketched in this commit, which adds a new extension trait that
any component can use to access common data.