Disable output quarantining.
We didn't finish implementing the client-side logic before realizing we needed
a better internal architecture, so the quarantining isn't usable at the moment,
and delegation/undelegation functionality therefore doesn't work.  In order to
avoid having to fix this simultaneously with the port to the new architecture,
temporarily disable quarantining, processing all notes immediately.  Once we
finish the port, we can reintroduce the quarantine logic simultaneously with
the required changes to the client protocol and to the (new) architecture.