Unlink the JMT from the legacy Postgres state
This commit fixes a problem where the `chain_params` key of the JMT was
mysteriously unavailable, even though it had been written, causing a crash.
The problem was that although we started using the JMT in the new component
code, we didn't remove its use from the legacy Postgres state.  This caused the
legacy postgres code to clobber the tree roots, so although the tree roots were
actually written to the backing store, they were inaccessible, because the
roots had been overwritten with ones that didn't point to those leaves.

In the short term, the fix for this is to remove the JMT from the legacy state
code, so that there is only one writer.

In the long term, we should probably try to have the `Storage` impl return an
error if it is about to clobber existing data; since the JMT is an append-only
data structure, we should never have the backing store overwriting existing
key/value pairs.  Maybe there is an easy way to do this in RocksDB.