pd: unify note minting between ShieldedPool and legacy code
Currently, all new transactions fail verification in the `ShieldedPool`
component, because the `ShieldedPool` NCT and the legacy NCT de-sync and have
different commitment roots.  There are two reasons for this:

- the legacy code used a different (worse) mechanism for deterministically deriving the blinding factors for minted notes;
- the new code does not mint reward notes on epoch transitions.

To fix the first point, this commit backports the hash-the-NCT-root method to
the legacy code.

The second point is unfixed, so following this commit, the tree state will
continue to de-sync following the first epoch transition.  To fix it, we need
to mint reward notes in the `ShieldedPool` component.