pd: factor out Staking JMT accesses into a View trait
This is the same design as #550, but with a more evocative name ("view", since
it's a view of the data in the tree).

Having all of the state accesses be factored into a trait that defines what
keys it uses means that we can have all of the reading/writing done in one
place, and next to each other, which makes tracking what tree keys are used
much easier (we just have to look at two adjacent functions, and don't need to
check anything else or replicate the keys across the source tree).