stake: add a DelegationChanges struct and proto
In the previous staking code, we relied on the existence of
```
Reader::delegation_changes() -> BTreeMap<IdentityKey, i64>
```
that aggregated the delegation changes over an entire epoch at the database level.

This won't work going forward, because we won't have access to the individual
(un)delegation values, they'll be protected with flow encryption.  So we need a
way to record and then later extract a list of (un)delegation actions, which
the `DelegationChanges` struct now provides.