"Correctly" report AppHash in Info messages
We now report the AppHash from the RocksDB-backed JMT code, since we don't have
a JMT implementation in the legacy implementation any more.  This means that
the node will fail to restart, because (until this commit) it would query the
Postgres instance for the app hash, returning a dummy value.

Also this revealed that the height cache / notification channel thing never
even worked.  We need to trash all of this code ASAP