pd: add a new, Component-based mempool.
Closes #521

Currently, the new mempool is threaded through the old mempool, and its
processing is executed on a newly spawned task, with the results printed to a
tracing event and then forgotten.  This lets us compare validation results
against the old code (they're currently different, because of missing reward
notes causing NCT root divergence), and then drop the old mempool when we
migrate.