pd: use the NCT position rather than the root for blinding factor uniqueness
The existing implementation of `mint_note` is very slow, because the
`incrementalmerkletree` implementation isn't particularly fast at computing the
tree root, and we compute it for every genesis allocation.

As noted in the comment there, the point is just to achieve uniqueness:

> These notes are public, so we don't need a blinding factor for privacy,
> but since the note commitments are determined by the note contents, we
> need to have unique (deterministic) blinding factors for each note, so they
> cannot collide.
>
> Hashing the current NCT root is sufficient, since it will change every time
> we insert a new note.

Instead of hashing the root, we could also hash the position, which will also
change every time we insert a new note.  This isn't as good, since we're only
binding to the note count, rather than the tree state, but seems useful to do
at least for now, since waiting half an hour or more to process the initial app
state with all of the Discord faucet addresses is sad.