chain: add additional parameters for downtime-based slashing
Along the way, this also removes options from the `pd generate-testnet`
command, replacing them with an invocation of `Default::default()`, because
that seems more scalable long-term than trying to replicate every possible
parameter as a command-line option.  (It is a little annoying that if someone
wanted to set a different config, they would need to replicate it across all of
the generated configs, but a better fix for that would be to allow the
`generate-testnet` command to take a config as an option).