initial tokio-console integration
This commit adds initial support for the [Tokio Console], using the
basic [`console_subscriber::init`] API.  This should not interfere with
the existing logging configuration using `tracing`, as
`console_subscriber::init` is essentially a superset of
`tracing_subscriber::fmt::init`.

The console endpoint is served on the default address,
http://127.0.0.1:6669. Future commits will improve the console
instrumentation by adding task names, configure the console endpoint
using `pd`'s command line arguments, and integrate Tokio Console with
the existing Grafana setup, but this is a working starting point.

[Tokio Console]: https://docs.rs/tokio-console/latest/tokio_console/
[`console_subscriber::init`]: https://docs.rs/console-subscriber/latest/console_subscriber/fn.init.html