pd: fix vote tracking in staking component
The first attempt at implementing this functionality in #678 ignored the
`signed_last_block` field that records whether or not a validator actually
voted.  This commit uses it, and testing locally shows that it records missed
blocks (upload a validator definition with a random consensus key, delegate to
it, watch its missed block counter increase...).