pd: make the BeginBlock span smaller
It's not really useful to include the hash in the BeginBlock span, because the
other spans don't include it, so we don't have a way to filter all of the
processing of a block by its hash, and because Tendermint has finality, so it's
just as easy to identify a block by height as by hash.

Also, use a plain integer value in the span for consistency with the other messages.