proto: rework wallet protocol
This brings the wallet protocol in line with the design in #587, where we limit
the responsibilities of the wallet service to just sync and scan the public
state, leaving all of the logic of transaction creation to the client and
transaction signing to the (future) custody service.

In the future, this RPC could be extended with other parts of private chain
state (e.g., memo contents) that we don't currently download.

Closes #614