storage: replace State Mutex with RwLock
This commit changes the `State` typedef to use an `RwLock` instead of a
`Mutex`.  This allows concurrent read access to a shared `State`, which means
(because of the split between validation and execution in the `Component`
trait) that we can execute the `check_tx_stateful` methods of every `Component`
concurrently, even in the serial `consensus::Worker`.