crypto: rename `transaction::output::Body` to `crypto::NotePayload`
We use this structure in multiple places in the codebase, and it's
self-contained (being the minimal data payload needed to identify and scan a
new note).  So moving it into the `crypto` crate makes sense.

More importantly, it frees up the name `OutputBody` for other use.  In order to
have a useful signing/custody protocol, we need to classify transaction data
into *authorizing data* (which describes the actions that signatures authorize)
and *effecting data* (which effects those actions, but is not necessary to
authorize, e.g., proof data).  Then we'll want to have a protocol that allows
us to bundle up only the authorizing data and send it to a signer, who signs an
authorization hash (a renamed sighash) of the authorization data.  However,
this means that we want to pull apart the proto structures into chunks, where
each chunk has only authorizing /or/ effecting data.  To do /that/, it'd be
useful to be able to describe an `OutputBody`, which the `NotePayload` isn't
really.

Since the `NotePayload` is used much more widely than the other transaction
structures, splitting this out as a pre-change means that we can increase the
isolation of the remaining transaction structure refactoring from the rest of
the codebase.