pd: ensure components have shared state
This commit fixes a bug introduced in #796, which removed the `new` method from
the `Component` trait, in order to be able to "sideload" a copy of the note
commitment tree kept in a sidecar part of the underlying storage, rather than
in the (merklized) application state.

In that change, the `ShieldedPool` component was changed to take a raw
`Storage`, rather than a `State`, so that it would be able to read the NCT out
on its own.  But it's important that all of the components use the *same*
shared state, rather than each creating their own fork off of the version
currently in persistent storage.

Co-Authored-By: Ava Howell <ava@penumbra.zone>