transaction: split "authorizing" and "effecting" data
This commit changes the definition of the transaction structures and protos to
split out "authorizing data", data which describes what actions the transaction
does, which should be fed into an authorization hash for signing, and
"effecting data", which effects the actions described by the authorizing data,
but does not itself describe them, and which the authorization does not need to
be bound to.

This distinction is now consistently mapped to the `Foo`/`FooBody` distinction:
`FooBody` contains the authorizing data for `Foo`, and `Foo` contains both a
`FooBody` as well as any effecting data.

In particular, we move the anchor and the ZK proofs out of the `Body`
structures and consider the effecting data.  This allows a custodian to
authorize a transaction without restricting when the prover can create the
transaction proof.

Otherwise, if the anchor were included in the authorization hash, the prover
would be forced to use whatever anchor was available at the point that they
requested signing, and this creates a side channel that allows a third party to
learn information about what kind of custody was used for transaction
authorization.  For instance, threshold signing with offline keys necessarily
has high latency, which would force a prover to use an anchor that would be
hours or days out of date, distinguishing it from other transactions that used
recent anchors.